+++ src/app/app.module.ts
@@ -21,6 +21,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { SnackbarComponent } from './components/snackbar/snackbar.component';
 import { MatecuAlertBoxComponent, MatecuSpinnerComponent, MatecuTopbarActionsComponent, MatecuTopbarBodyComponent, MatecuTopbarFabComponent, MatecuTopbarLayoutComponent, MatecuTopbarLayoutModule, MatecuTopbarSearchComponent, MatecuTopbarTitleComponent } from '../../projects/angular-matecu/src/public-api';
 import { TopbarNextComponent } from './components/topbar-next/topbar-next.component';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 @NgModule({
   declarations: [
@@ -56,6 +57,7 @@ import { TopbarNextComponent } from './components/topbar-next/topbar-next.compon
     MatRadioModule,
     MatCheckboxModule,
     MatFormFieldModule,
+    MatSnackBarModule
   ],
   providers: [],
   bootstrap: [AppComponent],