+++ projects/angular-matecu/CHANGELOG.md
@@ -10,3 +10,4 @@ Versión 3.0
 - Se quito el boton de navegación pricipal, input actionMenu y output clickActionMenu fueron removidos
 - Se elimino componente matecu-topbar-actions queda en su lugar matecu-topbar-action que se puede utilizar en multiple ocaciones
 - Es obligatorio agrupar todos los elementos del topbar excepto matecu-topbar-body dentro del elemento matecu-topbar-header-row
+- Es obligatorio importar MatSnackBarModule para usar el componente matecu-alert-snack-bar